requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.2
pytz==2019.3
sqlparse==0.3.1
